+++ src/routes/userRoutes.js
@@ -1,23 +1,17 @@
 const express = require("express");
+const { userMiddleware } = require("../middleware/authMiddleware");
 const {
   createUser,
   loginUser,
-  myProfile,
-  updateProfile,
-  //getUser,
-  //getParticularUser,
-  //updateUser,
-  //deleteUser,
-} = require("../controllers/userController");
-const { userMiddleware } = require("../middleware/authMiddleware");
+  logOut,
+} = require("../controllers/authController");
+const { myProfile, updateProfile } = require("../controllers/userController");
 const userRouter = express.Router();
 
 userRouter.post("/Signup", createUser);
 userRouter.post("/Login", loginUser);
-userRouter.get("/myProfile/:id", userMiddleware, myProfile);
-userRouter.put("/updateProfile/:id", updateProfile);
-// userRouter.get("/users", getUser);
-// userRouter.get("/getParticularUser/:id", getParticularUser);
-//userRouter.put("/updateUser/:id", updateUser);
-//userRouter.delete("/deleteUser/:id", deleteUser);
+userRouter.get("/myProfile", userMiddleware, myProfile);
+userRouter.put("/updateProfile", userMiddleware, updateProfile);
+userRouter.post("/logout", userMiddleware, logOut);
+
 module.exports = userRouter;
